+++ DREAM Lab Frontend/src/pages/HomePage/components/Detalles/components/Etiqueta/Etiqueta.css
